pyproject.toml
[project]
name = "agentic-ai-course"
version = "0.1.0"
description = "Hands-on agentic AI course: LangGraph, OpenAI Agents SDK, and CrewAI, compared via a shared workflow-automation use case."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }

# Core dependencies - needed for Module 0 and the shared utilities.
# Keep this list minimal on purpose; module-specific libs live in optional groups.
dependencies = [
    "openai>=1.54.0",
    "python-dotenv>=1.0.1",
    "pydantic>=2.9.0",
    "rich>=13.9.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
# Module 1 - LangGraph
langgraph = [
    "langgraph>=0.2.50",
    "langgraph-checkpoint-sqlite>=2.0.0",
    "langchain-openai>=0.2.5",
    "langchain-core>=0.3.15",
]

# Module 2 - OpenAI Agents SDK
openai-agents = [
    "openai-agents>=0.0.7",
]

# Module 3 - CrewAI
crewai = [
    "crewai>=0.80.0",
    "crewai-tools>=0.15.0",
]

# Module 4 - Production (tracing, evals, safety)
production = [
    "langfuse>=2.50.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
]

# Module 5 - Capstone (real Google APIs + CLI)
capstone = [
    "google-api-python-client>=2.140.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.0",
    "typer>=0.12.0",
    "sqlalchemy>=2.0.30",
]

# Install everything (handy for development / CI):
#   pip install -e ".[all]"
all = [
    "agentic-ai-course[langgraph,openai-agents,crewai,production,capstone]",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["shared*"]
exclude = ["*.tests", "*.tests.*"]
